third_party/blink/web_tests/virtual/css-sizing-keyword-animation/external/wpt/css/css-sizing/animation/width-composition-expected.txt

This is a testharness.js-based test.
[FAIL] Compositing: property <width> underlying [100px] from add [100px] to add [auto] at (-0.3) should be [200px]
  assert_equals: expected "200px " but got "0px "
[FAIL] Compositing: property <width> underlying [100px] from add [100px] to add [auto] at (0.5) should be [auto]
  assert_equals: expected "784px " but got "542px "
[FAIL] Compositing: property <width> underlying [100px] from add [100px] to add [auto] at (1) should be [auto]
  assert_equals: expected "784px " but got "884px "
[FAIL] Compositing: property <width> underlying [100px] from add [100px] to add [auto] at (1.5) should be [auto]
  assert_equals: expected "784px " but got "1226px "
Harness: the test ran to completion.
